refactor: Enum 패키지 분리
