+++ batch-service/src/main/java/com/eatpizzaquickly/batchservice/coupon/batch/FeignUserCouponReader.java
@@ -6,12 +6,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.item.ItemReader;
 import org.springframework.data.domain.Page;
-
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.Collections;
 import java.util.Iterator;
-import java.util.List;
 
 @Slf4j
 @RequiredArgsConstructor
