+++ concert-service/src/main/java/com/eatpizzaquickly/concertservice/repository/ConcertRepository.java
@@ -1,6 +1,6 @@
 package com.eatpizzaquickly.concertservice.repository;
 
-import com.eatpizzaquickly.concertservice.entity.Category;
+import com.eatpizzaquickly.concertservice.enums.Category;
 import com.eatpizzaquickly.concertservice.entity.Concert;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
