--- coupon-service/src/main/java/com/eatpizzaquickly/couponservice/entity/CouponType.java
@@ -1,5 +1,0 @@
-package com.eatpizzaquickly.couponservice.entity;
-
-public enum CouponType {
-    ALL,LIMIT
-}
