+++ coupon-service/src/main/java/com/eatpizzaquickly/couponservice/enums/CouponType.java
@@ -1,0 +1,5 @@
+package com.eatpizzaquickly.couponservice.enums;
+
+public enum CouponType {
+    ALL,LIMIT
+}
