+++ reservation-service/src/main/java/com/eatpizzaquickly/reservationservice/common/enums/ReservationStatus.java
@@ -1,0 +1,5 @@
+package com.eatpizzaquickly.reservationservice.common.enums;
+
+public enum ReservationStatus {
+    PENDING,CONFIRMED,CANCELED
+}
