+++ reservation-service/src/main/java/com/eatpizzaquickly/reservationservice/payment/dto/PaymentRequestDto.java
@@ -1,7 +1,7 @@
 package com.eatpizzaquickly.reservationservice.payment.dto;
 
-import com.eatpizzaquickly.reservationservice.payment.entity.PayStatus;
-import com.eatpizzaquickly.reservationservice.payment.entity.SettlementStatus;
+import com.eatpizzaquickly.reservationservice.common.enums.PayStatus;
+import com.eatpizzaquickly.reservationservice.common.enums.SettlementStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
