+++ reservation-service/src/main/java/com/eatpizzaquickly/reservationservice/payment/service/PaymentService.java
@@ -12,15 +12,15 @@ import com.eatpizzaquickly.reservationservice.payment.dto.response.PaymentRespon
 import com.eatpizzaquickly.reservationservice.payment.dto.response.TossPaymentResponse;
 import com.eatpizzaquickly.reservationservice.payment.dto.response.UserResponseDto;
 import com.eatpizzaquickly.reservationservice.payment.dto.response.*;
-import com.eatpizzaquickly.reservationservice.payment.entity.PayMethod;
-import com.eatpizzaquickly.reservationservice.payment.entity.PayStatus;
+import com.eatpizzaquickly.reservationservice.common.enums.PayMethod;
+import com.eatpizzaquickly.reservationservice.common.enums.PayStatus;
 import com.eatpizzaquickly.reservationservice.payment.entity.Payment;
-import com.eatpizzaquickly.reservationservice.payment.entity.SettlementStatus;
+import com.eatpizzaquickly.reservationservice.common.enums.SettlementStatus;
 import com.eatpizzaquickly.reservationservice.payment.exception.*;
 import com.eatpizzaquickly.reservationservice.payment.kafka.PaymentEventProducer;
 import com.eatpizzaquickly.reservationservice.payment.repository.PaymentRepository;
 import com.eatpizzaquickly.reservationservice.reservation.entity.Reservation;
-import com.eatpizzaquickly.reservationservice.reservation.entity.ReservationStatus;
+import com.eatpizzaquickly.reservationservice.common.enums.ReservationStatus;
 import com.eatpizzaquickly.reservationservice.reservation.repository.ReservationRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
