--- reservation-service/src/main/java/com/eatpizzaquickly/reservationservice/reservation/entity/ReservationStatus.java
@@ -1,5 +1,0 @@
-package com.eatpizzaquickly.reservationservice.reservation.entity;
-
-public enum ReservationStatus {
-    PENDING,CONFIRMED,CANCELED
-}
