+++ user-service/src/main/java/com/eatpizzaquickly/userservice/service/UserService.java
@@ -9,7 +9,7 @@ import com.eatpizzaquickly.userservice.dto.UserRequestDto;
 import com.eatpizzaquickly.userservice.dto.UserResponseDto;
 import com.eatpizzaquickly.userservice.entity.HostBalance;
 import com.eatpizzaquickly.userservice.entity.User;
-import com.eatpizzaquickly.userservice.entity.UserRole;
+import com.eatpizzaquickly.userservice.enums.UserRole;
 import com.eatpizzaquickly.userservice.exception.*;
 import com.eatpizzaquickly.userservice.repository.HostBalanceRepository;
 import com.eatpizzaquickly.userservice.repository.UserRepository;
